refactor(manager): extract Instance struct and types to instance.go (#33)
Split main.go (370 lines) into 2 files:
- main.go: New() constructor, safeCallHandler, defaultSimState helpers
- instance.go: Instance struct definition + all handler entry types

No behavior changes — pure file reorganization.
